refactor: context와 provider 기능별 폴더분리
