+++ src/contexts/index.ts
@@ -1,7 +1,9 @@
-export { ThemeProvider, useThemeContext } from "./ThemeContext";
-export { UserProvider, useUserContext } from "./UserContext";
-export {
-  NotificationProvider,
-  useNotificationContext,
-} from "./NotificationContext";
-export { UserItemProvider, useUserItemContext } from "./UserItemContext";
+export { useThemeContext } from "./ThemeContext";
+export { useUserContext } from "./UserContext";
+export { useNotificationContext } from "./NotificationContext";
+export { useUserItemContext } from "./UserItemContext";
+
+export { ThemeProvider } from "./ThemeContext/ThemeProvider";
+export { UserProvider } from "./UserContext/UserProvider";
+export { NotificationProvider } from "./NotificationContext/NotificationProvider";
+export { UserItemProvider } from "./UserItemContext/UserItemProvider";
